refactor: refactor flag to application/flagparser
